move section extensions
