+++ src/libs/extension-api/controller/extension-api-initializer.test.ts
@@ -6,7 +6,8 @@ import { UmbControllerBase } from '@umbraco-cms/backoffice/class-api';
 import { UmbControllerHostElementMixin } from '@umbraco-cms/backoffice/controller-api';
 import type { UmbControllerHostElement, UmbControllerHost } from '@umbraco-cms/backoffice/controller-api';
 import { customElement, html } from '@umbraco-cms/backoffice/external/lit';
-import { type ManifestSection, UmbSwitchCondition } from '@umbraco-cms/backoffice/extension-registry';
+import { UmbSwitchCondition } from '@umbraco-cms/backoffice/extension-registry';
+import type { ManifestSection } from '@umbraco-cms/backoffice/section';
 
 @customElement('umb-test-controller-host')
 // Element is used in tests
