+++ src/packages/core/extension-registry/interfaces/index.ts
@@ -4,6 +4,3 @@ export * from './file-upload-preview.interface.js';
 export * from './menu-item-element.interface.js';
 export * from './modal-extension-element.interface.js';
 export * from './property-editor-ui-element.interface.js';
-export * from './section-element.interface.js';
-export * from './section-sidebar-app-element.interface.js';
-export * from './section-view-element.interface.js';
