+++ src/packages/core/section/extensions/index.ts
@@ -8,6 +8,8 @@ export type * from './section-sidebar-app.extension.js';
 export type * from './section-view.extension.js';
 export type * from './section.extension.js';
 
+export type * from './types.js';
+
 type UmbSectionExtensions =
 	| ManifestSection
 	| ManifestSectionRoute
