+++ src/packages/core/section/extensions/section-element.interface.ts
@@ -1,4 +1,4 @@
-import type { ManifestSection } from '../models/index.js';
+import type { ManifestSection } from './section.extension.js';
 
 export interface UmbSectionElement extends HTMLElement {
 	manifest?: ManifestSection;
