+++ src/packages/core/section/extensions/section-sidebar-app-element.interface.ts
@@ -1,4 +1,4 @@
-import type { ManifestSectionSidebarApp } from '../models/index.js';
+import type { ManifestSectionSidebarApp } from './section-sidebar-app.extension.js';
 
 export interface UmbSectionSidebarAppElement extends HTMLElement {
 	manifest?: ManifestSectionSidebarApp;
