+++ src/packages/core/section/extensions/types.ts
@@ -1,0 +1,3 @@
+export type * from './section-element.interface.js';
+export type * from './section-sidebar-app-element.interface.js';
+export type * from './section-view-element.interface.js';
