+++ src/packages/core/section/repository/item/types.ts
@@ -1,4 +1,4 @@
-import type { ManifestSection } from '@umbraco-cms/backoffice/extension-registry';
+import type { ManifestSection } from '@umbraco-cms/backoffice/section';
 
 export interface UmbSectionItemModel extends ManifestSection {
 	unique: string;
