+++ src/packages/core/section/types.ts
@@ -1,1 +1,2 @@
 export type * from './conditions/types.js';
+export type * from './extensions/index.js';
